Rename AbstractAnnotationDriverTestCas to AttributeDriverTest
