+++ tests/Tests/Mapping/AttributeDriverTest.php
@@ -4,15 +4,201 @@ declare(strict_types=1);
 
 namespace Doctrine\ODM\MongoDB\Tests\Mapping;
 
+use Doctrine\ODM\MongoDB\Mapping\Annotations as ODM;
+use Doctrine\ODM\MongoDB\Mapping\ClassMetadata;
 use Doctrine\ODM\MongoDB\Mapping\Driver\AttributeDriver;
+use Doctrine\ODM\MongoDB\Mapping\MappingException;
 use Doctrine\Persistence\Mapping\Driver\FileClassLocator;
-use Doctrine\Persistence\Mapping\Driver\MappingDriver;
+use Documents\CmsUser;
+use Generator;
+use PHPUnit\Framework\Attributes\DataProvider;
+use stdClass;
 
 use function class_exists;
 
-class AttributeDriverTest extends AbstractAnnotationDriverTestCase
+class AttributeDriverTest extends AbstractMappingDriverTestCase
 {
-    protected static function loadDriver(array $paths = []): MappingDriver
+    public function testFieldInheritance(): void
+    {
+        // @TODO: This can be a generic test for all drivers
+        $super  = $this->dm->getClassMetadata(AttributeDriverTestSuper::class);
+        $parent = $this->dm->getClassMetadata(AttributeDriverTestParent::class);
+        $child  = $this->dm->getClassMetadata(AttributeDriverTestChild::class);
+
+        self::assertFalse($super->hasField('id'), 'MappedSuperclass does not register its own public field');
+        self::assertFalse($super->hasField('protected'), 'MappedSuperclass does not register its own protected field');
+        self::assertTrue($super->hasField('private'), 'MappedSuperclass does register its own private field');
+        self::assertFalse($super->isInheritedField('private'), 'MappedSuperclass does not inherit its own field');
+        self::assertFalse($super->hasField('foo'), 'MappedSuperclass does not have field declared in child Document');
+        self::assertFalse($super->hasField('bar'), 'MappedSuperclass does not have field declared in grandchild Document');
+
+        self::assertArrayNotHasKey('declared', $super->fieldMappings['private'], 'MappedSuperclass does not track "declared" for non-inherited field');
+
+        self::assertTrue($parent->hasField('id'), 'Document does have public field from MappedSuperclass parent');
+        self::assertFalse($parent->isInheritedField('id'), 'Document does not inherit public field from MappedSuperclass parent');
+        self::assertTrue($parent->hasField('protected'), 'Document does have protected field from MappedSuperclass parent');
+        self::assertFalse($parent->isInheritedField('protected'), 'Document does not inherit protected field from MappedSuperclass parent');
+        self::assertTrue($parent->hasField('private'), 'Document does have private field from MappedSuperclass parent');
+        /* MappedSuperclass fields are never considered "inherited", but the
+         * field is still considered "declared" in the MappedSuperclass, since
+         * we need its ReflectionProperty to access it. This is a bit weird.
+         */
+        self::assertFalse($parent->isInheritedField('private'), 'Document does not inherit private field from MappedSuperclass parent');
+        self::assertTrue($parent->hasField('foo'), 'Document does register its own public field');
+        self::assertFalse($parent->isInheritedField('foo'), 'Document does not inherit its own field');
+        self::assertFalse($parent->hasField('bar'), 'Document does not have field declared in child Document');
+
+        self::assertArrayNotHasKey('declared', $parent->fieldMappings['id'], 'Document does not track "declared" for non-inherited public field from MappedSuperclass parent');
+        self::assertArrayNotHasKey('declared', $parent->fieldMappings['protected'], 'Document does not track "declared" for non-inherited protected field from MappedSuperclass parent');
+        self::assertEquals(AttributeDriverTestSuper::class, $parent->fieldMappings['private']['declared'], 'Non-inherited private field from MappedSuperclass parent is declared in MappedSuperclass parent');
+        self::assertArrayNotHasKey('declared', $parent->fieldMappings['foo'], 'Document does not track "declared" for its own public field');
+
+        self::assertTrue($child->hasField('id'), 'Document does have public field from MappedSuperclass grandparent');
+        self::assertTrue($child->isInheritedField('id'), 'Document does inherit public field from MappedSuperclass grandparent');
+        self::assertTrue($child->hasField('protected'), 'Document does have protected field from MappedSuperclass grandparent');
+        self::assertTrue($child->isInheritedField('protected'), 'Document does inherit protected field from MappedSuperclass grandparent');
+        self::assertTrue($child->hasField('private'), 'Document does have private field from MappedSuperclass grandparent');
+        self::assertTrue($child->isInheritedField('private'), 'Document does inherit private field from MappedSuperclass grandparent');
+        self::assertTrue($child->hasField('foo'), 'Document does have public field from Document parent');
+        self::assertTrue($child->isInheritedField('foo'), 'Document field declared in Document parent is inherited');
+        self::assertTrue($child->hasField('bar'), 'Document does register its own public field');
+        self::assertFalse($child->isInheritedField('bar'), 'Document does not inherit its own field');
+
+        self::assertEquals(AttributeDriverTestParent::class, $child->fieldMappings['id']['declared'], 'Inherited public field from MappedSuperclass grandparent is declared in Document parent');
+        self::assertEquals(AttributeDriverTestParent::class, $child->fieldMappings['protected']['declared'], 'Inherited protected field from MappedSuperclass grandparent is declared in Document parent');
+        self::assertEquals(AttributeDriverTestSuper::class, $child->fieldMappings['private']['declared'], 'Inherited private field from MappedSuperclass grandparent is declared in MappedSuperclass grandparent');
+        self::assertEquals(AttributeDriverTestParent::class, $child->fieldMappings['foo']['declared'], 'Inherited public field from Document parent is declared in Document parent');
+    }
+
+    public function testLoadMetadataForNonDocumentThrowsException(): void
+    {
+        $cm              = new ClassMetadata('stdClass');
+        $attributeDriver = static::loadDriver();
+
+        $this->expectException(MappingException::class);
+        $attributeDriver->loadMetadataForClass('stdClass', $cm);
+    }
+
+    public function testColumnWithMissingTypeDefaultsToString(): void
+    {
+        $cm              = new ClassMetadata(ColumnWithoutType::class);
+        $attributeDriver = static::loadDriver();
+
+        $attributeDriver->loadMetadataForClass(stdClass::class, $cm);
+        self::assertEquals('id', $cm->fieldMappings['id']['type']);
+    }
+
+    public function testGetAllClassNamesReturnsAlreadyLoadedClassesIfAppropriate(): void
+    {
+        self::assertTrue(class_exists(CmsUser::class), 'Pre-load the class');
+        $attributeDriver = $this->loadDriverForCMSDocuments();
+        $classes         = $attributeDriver->getAllClassNames();
+
+        self::assertContains(CmsUser::class, $classes);
+    }
+
+    public function testGetClassNamesReturnsOnlyTheAppropriateClasses(): void
+    {
+        $extraneousClassName = ColumnWithoutType::class;
+
+        $attributeDriver = $this->loadDriverForCMSDocuments();
+        $classes         = $attributeDriver->getAllClassNames();
+
+        self::assertNotContains($extraneousClassName, $classes);
+    }
+
+    public function testEmbeddedClassCantHaveShardKey(): void
+    {
+        $this->expectException(MappingException::class);
+        $this->expectExceptionMessage('Embedded document can\'t have shard key');
+        $this->dm->getClassMetadata(AttributeDriverEmbeddedWithShardKey::class);
+    }
+
+    public function testDocumentAttributeCanSpecifyWriteConcern(): void
+    {
+        $cm = $this->dm->getClassMetadata(AttributeDriverTestWriteConcernMajority::class);
+        self::assertSame('majority', $cm->writeConcern);
+
+        $cm = $this->dm->getClassMetadata(AttributeDriverTestWriteConcernUnacknowledged::class);
+        self::assertSame(0, $cm->writeConcern);
+
+        $cm = $this->dm->getClassMetadata(ColumnWithoutType::class);
+        self::assertNull($cm->writeConcern);
+    }
+
+    #[DataProvider('provideClassCanBeMappedByOneAbstractDocument')]
+    public function testClassCanBeMappedByOneAbstractDocument(object $wrong, string $messageRegExp): void
+    {
+        $this->expectException(MappingException::class);
+        $this->expectExceptionMessageMatches($messageRegExp);
+
+        $cm     = new ClassMetadata($wrong::class);
+        $driver = static::loadDriver();
+
+        $driver->loadMetadataForClass($wrong::class, $cm);
+    }
+
+    public static function provideClassCanBeMappedByOneAbstractDocument(): ?Generator
+    {
+        yield [
+            new #[ODM\Document]
+                #[ODM\EmbeddedDocument]
+            class () {
+            },
+            '/as EmbeddedDocument because it was already mapped as Document\.$/',
+        ];
+
+        yield [
+            new #[ODM\Document]
+                #[ODM\File]
+            class () {
+            },
+            '/as File because it was already mapped as Document\.$/',
+        ];
+
+        yield [
+            new #[ODM\Document]
+                #[ODM\QueryResultDocument]
+            class () {
+            },
+            '/as QueryResultDocument because it was already mapped as Document\.$/',
+        ];
+
+        yield [
+            new #[ODM\Document]
+                #[ODM\View]
+            class () {
+            },
+            '/as View because it was already mapped as Document\.$/',
+        ];
+
+        yield [
+            new #[ODM\Document]
+                #[ODM\MappedSuperclass]
+            class () {
+            },
+            '/as MappedSuperclass because it was already mapped as Document\.$/',
+        ];
+
+        yield [
+            new #[ODM\MappedSuperclass]
+                #[ODM\Document]
+            class () {
+            },
+            '/as Document because it was already mapped as MappedSuperclass\.$/',
+        ];
+    }
+
+    public function testWrongValueForValidationValidatorShouldThrowException(): void
+    {
+        $driver        = static::loadDriver();
+        $classMetadata = new ClassMetadata(WrongValueForValidationValidator::class);
+        $this->expectException(MappingException::class);
+        $this->expectExceptionMessage('The following schema validation error occurred while parsing the "validator" property of the "Doctrine\ODM\MongoDB\Tests\Mapping\WrongValueForValidationValidator" class: "Got parse error at "w", position 0: "SPECIAL_EXPECTED"" (code 0).');
+        $driver->loadMetadataForClass($classMetadata->name, $classMetadata);
+    }
+
+    protected static function loadDriver(array $paths = []): AttributeDriver
     {
         if (class_exists(FileClassLocator::class)) {
             $paths = FileClassLocator::createFromDirectories($paths);
@@ -20,4 +206,89 @@ class AttributeDriverTest extends AbstractAnnotationDriverTestCase
 
         return AttributeDriver::create($paths);
     }
+
+    protected function loadDriverForCMSDocuments(): AttributeDriver
+    {
+        return static::loadDriver([__DIR__ . '/../../Documents']);
+    }
+}
+
+
+#[ODM\Document]
+class ColumnWithoutType
+{
+    /** @var string|null */
+    #[ODM\Id]
+    public $id;
+}
+
+
+#[ODM\MappedSuperclass]
+class AttributeDriverTestSuper
+{
+    /** @var string|null */
+    #[ODM\Id]
+    public $id;
+
+    /** @var string|null */
+    #[ODM\Field(type: 'string')]
+    protected $protected;
+
+    /** @var string|null */
+    #[ODM\Field(type: 'string')]
+    private $private;
+}
+
+
+#[ODM\Document]
+class AttributeDriverTestParent extends AttributeDriverTestSuper
+{
+    /** @var string|null */
+    #[ODM\Field(type: 'string')]
+    public $foo;
+}
+
+
+#[ODM\Document]
+class AttributeDriverTestChild extends AttributeDriverTestParent
+{
+    /** @var string|null */
+    #[ODM\Field(type: 'string')]
+    public $bar;
+}
+
+#[ODM\EmbeddedDocument]
+#[ODM\ShardKey(keys: ['foo' => 'asc'])]
+class AttributeDriverEmbeddedWithShardKey
+{
+    /** @var string|null */
+    #[ODM\Field(type: 'string')]
+    public $foo;
+}
+
+
+#[ODM\Document(writeConcern: 'majority')]
+class AttributeDriverTestWriteConcernMajority
+{
+    /** @var string|null */
+    #[ODM\Id]
+    public $id;
+}
+
+
+#[ODM\Document(writeConcern: 0)]
+class AttributeDriverTestWriteConcernUnacknowledged
+{
+    /** @var string|null */
+    #[ODM\Id]
+    public $id;
+}
+
+
+#[ODM\Validation(validator: 'wrong')]
+class WrongValueForValidationValidator
+{
+    /** @var string|null */
+    #[ODM\Id]
+    public $id;
 }
